requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
flask-cors==6.0.2
google-generativeai==0.9.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
numpy==2.4.1
pandas==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
Werkzeug==3.1.5
